refactor(flight): split Terminal Event Detection into ParsekFlight.TerminalEvents.cs
